requirements.txt
# Monday.com BI Agent Dependencies

# Core dependencies
streamlit>=1.30.0
openai>=1.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: for future enhancements
# plotly>=5.18.0
# numpy>=1.24.0